moved topological features to 2nd assumption
