+++ examples/simple/time_series_forecasting/api_forecasting.py
@@ -5,15 +5,24 @@ import numpy as np
 import pandas as pd
 from matplotlib import pyplot as plt
 
-from examples.advanced.time_series_forecasting.multistep import TS_DATASETS
 from fedot import Fedot
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.pipelines.pipeline_builder import PipelineBuilder
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import TsForecastingParams, Task, TaskTypesEnum
+from fedot.core.utils import fedot_project_root
 
 logging.raiseExceptions = False
+
+_TS_EXAMPLES_DATA_PATH = fedot_project_root().joinpath('examples/data/ts')
+
+TS_DATASETS = {
+    'm4_daily': _TS_EXAMPLES_DATA_PATH.joinpath('M4Daily.csv'),
+    'm4_monthly': _TS_EXAMPLES_DATA_PATH.joinpath('M4Monthly.csv'),
+    'm4_quarterly': _TS_EXAMPLES_DATA_PATH.joinpath('M4Quarterly.csv'),
+    'm4_weekly': _TS_EXAMPLES_DATA_PATH.joinpath('M4Weekly.csv'),
+    'm4_yearly': _TS_EXAMPLES_DATA_PATH.joinpath('M4Yearly.csv')}
 
 
 def get_ts_data(dataset='m4_monthly', horizon: int = 30, m4_id=None, validation_blocks=None):
@@ -28,11 +37,7 @@ def get_ts_data(dataset='m4_monthly', horizon: int = 30, m4_id=None, validation_
     print(label)
     time_series = time_series[time_series['label'] == label]
 
-    if dataset not in ['australia']:
-        idx = pd.to_datetime(time_series['idx'].values)
-    else:
-        # non datetime indexes
-        idx = time_series['idx'].values
+    idx = time_series['idx'].values
 
     time_series = time_series['value'].values
     train_input = InputData(idx=idx,
@@ -80,4 +85,4 @@ def run_ts_forecasting_example(dataset='australia', horizon: int = 30, timeout: 
 
 
 if __name__ == '__main__':
-    run_ts_forecasting_example(dataset='m4_monthly', validation_blocks=None, horizon=14, timeout=2., visualization=True)
+    run_ts_forecasting_example(dataset='m4_monthly', horizon=14, timeout=2., visualization=True)
